+++ how-to/customize-workspace/client/src/framework/shapes/integrations-shapes.ts
@@ -6,6 +6,7 @@ import type {
 	HomeSearchResult
 } from "@openfin/workspace";
 import type { BrowserWindowModule, Page } from "@openfin/workspace-platform";
+import type { IShareCustomData } from "customize-workspace/share";
 import type { ModuleDefinition, ModuleHelpers, ModuleImplementation, ModuleList } from "./module-shapes";
 import type { TemplateHelpers } from "./template-shapes";
 
@@ -24,7 +25,7 @@ export interface IntegrationHelpers extends ModuleHelpers {
 	 * @param targetIdentity The optional target identity of the launch with.
 	 * @returns The launched view.
 	 */
-	launchView?(
+	launchView(
 		view: OpenFin.PlatformViewCreationOptions | string,
 		targetIdentity?: OpenFin.Identity
 	): Promise<OpenFin.View>;
@@ -35,25 +36,43 @@ export interface IntegrationHelpers extends ModuleHelpers {
 	 * @param bounds The optional bounds for the page.
 	 * @returns The window created.
 	 */
-	launchPage?(page: Page, bounds?: OpenFin.Bounds): Promise<BrowserWindowModule>;
+	launchPage(page: Page, bounds?: OpenFin.Bounds): Promise<BrowserWindowModule>;
 
 	/**
 	 * Launch a snapshot.
 	 * @param snapshotUrl The snapshot url
 	 */
-	launchSnapshot?(snapshotUrl: string): Promise<OpenFin.Identity[]>;
+	launchSnapshot(snapshotUrl: string): Promise<OpenFin.Identity[]>;
 
 	/**
 	 * Open a url with the browser.
 	 * @param url The url to open.
 	 */
-	openUrl?(url: string): Promise<void>;
+	openUrl(url: string): Promise<void>;
 
 	/**
 	 * Set the home search query.
 	 * @param query The query to set.
 	 */
-	setSearchQuery?(query: string): Promise<void>;
+	setSearchQuery(query: string): Promise<void>;
+
+	/**
+	 * Get the value of a condition.
+	 * @param conditionId The id of the condition.
+	 * @returns True if the condition is set.
+	 */
+	condition(conditionId: string): Promise<boolean>;
+
+	/**
+	 * Share data.
+	 * @param options The sharing options.
+	 */
+	share(options?: IShareCustomData): Promise<void>;
+
+	/**
+	 * Create a random UUID.
+	 */
+	randomUUID(): string;
 }
 
 /**
@@ -116,12 +135,17 @@ export interface IntegrationModule<O = unknown> extends ModuleImplementation<O, 
 	 * @param query The query to search for.
 	 * @param filters The filters to apply.
 	 * @param lastResponse The last search response used for updating existing results.
+	 * @param options Options for the search query.
 	 * @returns The list of results and new filters.
 	 */
 	getSearchResults?(
 		query: string,
 		filters: CLIFilter[],
-		lastResponse: HomeSearchListenerResponse
+		lastResponse: HomeSearchListenerResponse,
+		options: {
+			queryMinLength: number;
+			queryAgainst: string[];
+		}
 	): Promise<HomeSearchResponse>;
 
 	/**