+++ how-to/customize-workspace/client/src/framework/shapes/module-shapes.ts
@@ -78,6 +78,11 @@ export interface ModuleHelpers {
 	getCurrentThemeId(): Promise<string>;
 
 	/**
+	 * Get the current icon folder.
+	 */
+	getCurrentIconFolder(): Promise<string>;
+
+	/**
 	 * Get the current palette.
 	 */
 	getCurrentPalette(): Promise<CustomPaletteSet>;