+++ how-to/customize-workspace/CHANGELOG.md
@@ -17,6 +17,10 @@
 - Added opinionated support for fdc3.open (see [How To Add Open Support To Your App](./docs/how-to-add-open-support-to-your-app.md))
 - Updated example auth to give an example of how workspace can present entitled based data by letting you pick from two users and each user has a role (developer or sales) and the developer role has an apps feed that consists of developer related apps and has menu options that help developers (inspect view, window, platform etc) where as the sales role keeps the demo apps but filters out developer related menu options and apps. **npm run secondclient** will run the demo.
 - Moved the create app definition action logic from the app definition module (which has been deleted) into a developer action module which also has inspect menu actions.
+- Moved the workspaces home logic into the Workspaces integration provider
+- Moved the pages home logic into the Pages integration provider
+- Removed `enableWorkspaceIntegration` from `homeProvider` - enable/disable the Workspaces integration instead
+- Removed `enablePageIntegration` from `homeProvider` - enable/disable the Pages integration instead
 
 ## v9.2